+++ src/main/java/kitchenpos/application/order/OrderService.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.order;
 
 import java.util.ArrayList;
 import java.util.List;
