+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -26,6 +26,7 @@ import kitchenpos.application.fixture.OrderFixtureFactory;
 import kitchenpos.application.fixture.OrderLineItemFixtureFactory;
 import kitchenpos.application.fixture.OrderTableFixtureFactory;
 import kitchenpos.application.fixture.ProductFixtureFactory;
+import kitchenpos.application.order.OrderService;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.menu.MenuRepository;
